+++ website/docs/releases/release-notes.md
@@ -9,6 +9,7 @@ title: Release Notes
 - :boom: **BREAKING**: Make `provider` a required option in `Contract` constructor.
 - :boom: **BREAKING**: Remove deprecated "old" transaction builder (`contract.functions`).
 - :boom: **BREAKING**: No longer seed the MockNetworkProvider with any test UTXOs.
+- :sparkles: Add a configurable `vmTarget` option to `MockNetworkProvider`.
 - :hammer_and_wrench: Improve libauth template generation.
 
 ## v0.11.5